+++ src/main/kotlin/ememies/behaviour/AggressiveStrategy.kt
@@ -1,7 +1,8 @@
-package ememies
+package ememies.behaviour
 
-import graphics.model.Character
-import graphics.model.MapChecker
+import ememies.Mob
+import player.Character
+import utils.MapChecker
 import kotlin.math.absoluteValue
 
 class AggressiveStrategy: BehaviourStrategy() {
