+++ src/main/kotlin/graphics/map_loading/BadMapFileException.kt
@@ -1,4 +1,4 @@
-package graphics.model
+package graphics.map_loading
 
 /** thrown during map file reading
  * in case of file format error when meets unsupported symbol */
